+++ test/core_tests/version_test.cpp
@@ -6,6 +6,6 @@
 using namespace owl;
 
 TEST(Core_Version, base) {
-	EXPECT_STREQ("0.1.1", getVersionString().c_str());
-	EXPECT_EQ(0x00010100, getVersionCode());
+	EXPECT_STREQ("0.2.0", getVersionString().c_str());
+	EXPECT_EQ(0x00020000, getVersionCode());
 }